Updated post_grades() to send raw score and max points
It is more useful for course staff if the remote gradebook
is sent the student's raw score for an assignment and the
assignment contains the maximum points allowable.  edX now
sends raw score and maximum points allowed in its CSV file.
``post_grades() now explicitly creates the assignment, assigning
the maximum points value and then sends the data.

Fixes #36

@pdpinch